Replaced custom popover with what's in BitKit
