--- source/javascripts/components/ProductTour/ProductTooltip.scss
@@ -1,48 +1,0 @@
-.product-tooltip {
-    background-color: white;
-    border: 1px solid rgba(239, 235, 239, 0.5);
-    box-shadow: 0px 2px 24px rgba(0, 0, 0, 0.08);
-    border-radius: 8px;
-    max-width: 420px;
-    min-height: 212px;
-    padding: 24px;
-    z-index: 300;
-    position: absolute;
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-
-    .arrow-up {
-      width: 0;
-      height: 0;
-      border-left: 10px solid transparent;
-      border-right: 10px solid transparent;
-      border-bottom: 10px solid white;
-      position: absolute;
-      top: -10px;
-      left: 50%;
-    }
-
-    .arrow-left {
-      width: 0;
-      height: 0;
-      border-right: 10px solid white;
-      border-bottom: 10px solid transparent;
-      border-top: 10px solid transparent;
-      position: absolute;
-      left: -10px;
-      top: 50%;
-    }
-  }
-
-  .product-tooltip__footer {
-    display: flex;
-    justify-content: space-between;
-  }
-
-  .product-tooltip__navigation {
-    display: flex;
-    gap: 8px;
-  }
-
-
